refactor styles of NewFeedList
