+++ src/sidebar/NewFeedForm/NewFeedsList/NewFeedsList.tsx
@@ -1,45 +1,23 @@
-import styled from '@emotion/styled';
 import { Warning, Spinner, Check } from 'phosphor-react';
 
-import React, { Fragment, FunctionComponent } from 'react';
+import React, { FunctionComponent } from 'react';
 
-import { useAppSelector } from '../../store/hooks';
-import { UnreachableCaseError } from '../../utils/UnreachableCaseError';
-
-const List = styled.ul`
-    flex: 1;
-    padding-left: 12px;
-    border: 1px solid;
-    border-color: ${(props) => props.theme.colors.menuBorderColor};
-    border-radius: 4px;
-    margin-top: 6px;
-    overflow-y: scroll;
-`;
-
-const ListItem = styled.li`
-    padding: 4px;
-
-    list-style: none;
-
-    & > svg {
-        margin-right: 6px;
-        vertical-align: middle;
-    }
-`;
+import { useAppSelector } from '../../../store/hooks';
+import { UnreachableCaseError } from '../../../utils/UnreachableCaseError';
 
 interface Props {
     newFeedUrls: ReadonlyArray<string>;
 }
 
-const NewFeedsList: FunctionComponent<Props> = (props: Props) => {
+export const NewFeedsList: FunctionComponent<Props> = (props: Props) => {
     const feeds = useAppSelector((state) => state.feeds.feeds);
     const feedStatus = useAppSelector((state) => state.session.feedStatus);
 
     return (
-        <Fragment>
-            <label className="subscribe-view__section-heading">Recently added Feeds:</label>
+        <>
+            <label className="subscribe-view__section-heading">Recently added Feeds</label>
             {props.newFeedUrls.length > 0 && (
-                <List>
+                <ul className="subscribe-view__new-feeds-list">
                     {props.newFeedUrls.map((newFeedUrl) => {
                         const status = feedStatus.find((s) => s.url === newFeedUrl)?.status;
 
@@ -50,20 +28,20 @@ const NewFeedsList: FunctionComponent<Props> = (props: Props) => {
                         switch (status) {
                             case 'loading':
                                 return (
-                                    <ListItem key={newFeedUrl}>
+                                    <li className="subscribe-view__new-feeds-list-item" key={newFeedUrl}>
                                         <Spinner size={18} />
                                         loading...
-                                    </ListItem>
+                                    </li>
                                 );
 
                             case 'loaded': {
                                 const loadedFeed = feeds.find((f) => f.id === newFeedUrl);
                                 if (loadedFeed !== undefined) {
                                     return (
-                                        <ListItem key={newFeedUrl}>
+                                        <li className="subscribe-view__new-feeds-list-item" key={newFeedUrl}>
                                             <Check size={18} />
                                             {loadedFeed.title}
-                                        </ListItem>
+                                        </li>
                                     );
                                 } else {
                                     return <li>{newFeedUrl}</li>;
@@ -72,20 +50,17 @@ const NewFeedsList: FunctionComponent<Props> = (props: Props) => {
 
                             case 'error':
                                 return (
-                                    <ListItem key={newFeedUrl}>
+                                    <li className="subscribe-view__new-feeds-list-item" key={newFeedUrl}>
                                         <Warning size={18} /> {newFeedUrl}
-                                    </ListItem>
+                                    </li>
                                 );
 
                             default:
                                 throw new UnreachableCaseError(status);
                         }
                     })}
-                </List>
+                </ul>
             )}
-            <div></div>
-        </Fragment>
+        </>
     );
 };
-
-export default NewFeedsList;
